feat(intergration_test): add intergration test
