+++ tests/test_replicate.py
@@ -43,21 +43,6 @@ def test_multithread_google(get_file_from_manifest,monkeypatch):
     google.run()
     assert scripts.replicate.exec_google_copy.called == True
 
-@patch('scripts.replicate.get_fileinfo_list_from_manifest')
-def test_multithread_aws(get_file_from_manifest,monkeypatch,initialization):
-    monkeypatch.setattr(
-            'scripts.replicate.MODE', 'test')
-    get_file_from_manifest.return_value = gen_mock_manifest_data()
-    number_of_threads = 4
-    scripts.replicate.call_aws_copy = MagicMock()
-    aws = AWSBucketReplication({'from_bucket': 'from','to_bucket': 'to'},'test',number_of_threads)
-    aws.prepare()
-    aws.run()
-    assert scripts.replicate.call_aws_copy.called == True
-
-
-
-
 #patch('scripts.replicate.AMZBucketReplication.get_submitting_indexd_files')
 #patch('scripts.sync.AMZBucketReplication.get_file_from_uuid')
 #ef test_uuid_not_existed(get_index_uuid, get_submitting_files):
